Adding additional information from requestedtemp.json
Adding additional information from requestedtemp.json to post to backend
Azure for ML and analytics
